refactor: add native query operation support
Extends the executor's native operation system to support query operations for
reading blockchain state. This separates the concerns between state-modifying
calls and read-only queries.

Changes include:
- Add NativeQuery enum for read-only operations
- Create separate handlers for calls and queries
- Split RPC methods into call and query endpoints
- Refactor task handling to support both operation types
