+++ tee-worker/omni-executor/rpc-server/src/methods/submit_plain_request.rs
@@ -6,7 +6,7 @@ use crate::{
 	request::PlainRequest,
 	server::RpcContext,
 };
-use executor_core::native_operation::NativeCall;
+use executor_core::native_operation::{NativeOperation, NativeQuery};
 use executor_primitives::{
 	utils::hex::{FromHexPrefixed, ToHexPrefixed},
 	OmniAccountAuthType,
@@ -15,13 +15,13 @@ use jsonrpsee::{
 	types::{ErrorCode, ErrorObject},
 	RpcModule,
 };
-use native_task_handler::NativeTask;
+use native_task_handler::{NativeTask, NativeTaskOperation};
 use parentchain_rpc_client::{SubstrateRpcClient, SubstrateRpcClientFactory};
 use parity_scale_codec::Decode;
 use std::sync::Arc;
 use tokio::{runtime::Handle, sync::oneshot, task};
 
-pub fn register_submit_plain_request<
+pub fn register_submit_plain_requests<
 	Header: Send + Sync + 'static,
 	RpcClient: SubstrateRpcClient<Header> + Send + Sync + 'static,
 	RpcClientFactory: SubstrateRpcClientFactory<Header, RpcClient> + Send + Sync + 'static,
@@ -29,7 +29,7 @@ pub fn register_submit_plain_request<
 	module: &mut RpcModule<RpcContext<Header, RpcClient, RpcClientFactory>>,
 ) {
 	module
-		.register_async_method("native_submitPlainRequest", |params, ctx, _| async move {
+		.register_async_method("native_submitCallPlainRequest", |params, ctx, _| async move {
 			let Ok(hex_request) = params.one::<String>() else {
 				return Err(ErrorCode::ParseError.into());
 			};
@@ -46,8 +46,11 @@ pub fn register_submit_plain_request<
 				ErrorCode::InternalError
 			})??;
 			let (response_sender, response_receiver) = oneshot::channel();
-			let native_task = NativeTask { call: native_call, auth_type, response_sender };
-
+			let native_task = NativeTask {
+				operation: NativeTaskOperation::Call(native_call),
+				auth_type,
+				response_sender,
+			};
 			if ctx.native_task_sender.send(native_task).await.is_err() {
 				log::error!("Failed to send request to native call executor");
 				return Err(ErrorCode::InternalError.into());
@@ -60,19 +63,57 @@ pub fn register_submit_plain_request<
 				},
 			}
 		})
-		.expect("Failed to register native_submitPlainRequest method");
+		.expect("Failed to register native_submitCallPlainRequest method");
+
+	module
+		.register_async_method("native_submitQueryPlainRequest", |params, ctx, _| async move {
+			let Ok(hex_request) = params.one::<String>() else {
+				return Err(ErrorCode::ParseError.into());
+			};
+			let Ok(request) = PlainRequest::from_hex(&hex_request) else {
+				return Err(ErrorCode::ServerError(INVALID_PLAIN_REQUEST_CODE).into());
+			};
+			let join_handle = task::spawn_blocking({
+				let ctx = ctx.clone();
+				let plain_request = request.clone();
+				|| handle_plain_request(plain_request, ctx, Handle::current())
+			});
+			let (native_query, auth_type) = join_handle.await.map_err(|e| {
+				log::error!("Failed to handle Plain request: {:?}", e);
+				ErrorCode::InternalError
+			})??;
+			let (response_sender, response_receiver) = oneshot::channel();
+			let native_task = NativeTask {
+				operation: NativeTaskOperation::Query(native_query),
+				auth_type,
+				response_sender,
+			};
+			if ctx.native_task_sender.send(native_task).await.is_err() {
+				log::error!("Failed to send request to native call executor");
+				return Err(ErrorCode::InternalError.into());
+			}
+			match response_receiver.await {
+				Ok(response) => Ok::<String, ErrorObject>(response.to_hex()),
+				Err(e) => {
+					log::error!("Failed to receive response from native call handler: {:?}", e);
+					Err(ErrorCode::InternalError.into())
+				},
+			}
+		})
+		.expect("Failed to register native_submitQueryPlainRequest method");
 }
 
-fn handle_plain_request<
+pub fn handle_plain_request<
 	'a,
 	Header,
 	RpcClient: SubstrateRpcClient<Header>,
 	RpcClientFactory: SubstrateRpcClientFactory<Header, RpcClient>,
+	OP: NativeOperation,
 >(
 	request: PlainRequest,
 	ctx: Arc<RpcContext<Header, RpcClient, RpcClientFactory>>,
 	handle: Handle,
-) -> Result<(NativeCall, OmniAccountAuthType), ErrorObject<'a>> {
+) -> Result<(OP, OmniAccountAuthType), ErrorObject<'a>> {
 	if request.mrenclave != ctx.mrenclave {
 		return Err(ErrorCode::ServerError(INVALID_MRENCLAVE_CODE).into());
 	}
